Move "add" functionality into new file
